add classes player and game
